refactor: 서버 패키지를 레이어 기반으로 재구성
domain.memo.* → controller / service / repository / dto / domain / error

- controller: MemoController, HealthController
- service: MemoService
- repository: MemoRepository, MemoTombstoneRepository
- domain: Memo, MemoTombstone, MemoCategories, MemoStickyColor
- dto: 모든 Request/Response/DTO (HealthResponse 포함)
- error: 빈 패키지 (GlobalExceptionHandler 등 이후 추가 예정)
