+++ apps/server-spring/src/main/java/com/ainote/server/controller/MemoController.java
@@ -1,13 +1,14 @@
-package com.ainote.server.memo;
+package com.ainote.server.controller;
 
-import com.ainote.server.memo.dto.CreateMemoResponse;
-import com.ainote.server.memo.dto.DeleteMemoResponse;
-import com.ainote.server.memo.dto.GetMemoResponse;
-import com.ainote.server.memo.dto.ListDeletionsResponse;
-import com.ainote.server.memo.dto.ListMemosResponse;
-import com.ainote.server.memo.dto.MemoCreateRequest;
-import com.ainote.server.memo.dto.MemoUpsertRequest;
-import com.ainote.server.memo.dto.UpdateMemoResponse;
+import com.ainote.server.dto.CreateMemoResponse;
+import com.ainote.server.dto.DeleteMemoResponse;
+import com.ainote.server.dto.GetMemoResponse;
+import com.ainote.server.dto.ListDeletionsResponse;
+import com.ainote.server.dto.ListMemosResponse;
+import com.ainote.server.dto.MemoCreateRequest;
+import com.ainote.server.dto.MemoUpsertRequest;
+import com.ainote.server.dto.UpdateMemoResponse;
+import com.ainote.server.service.MemoService;
 import java.time.Instant;
 import java.util.Map;
 import org.springframework.http.HttpStatus;
