+++ apps/server-spring/src/main/java/com/ainote/server/domain/MemoStickyColor.java
@@ -1,4 +1,4 @@
-package com.ainote.server.memo;
+package com.ainote.server.domain;
 
 import java.util.Set;
