+++ apps/server-spring/src/main/java/com/ainote/server/domain/MemoTombstone.java
@@ -1,4 +1,4 @@
-package com.ainote.server.memo;
+package com.ainote.server.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
