+++ apps/server-spring/src/main/java/com/ainote/server/dto/GetMemoResponse.java
@@ -1,4 +1,4 @@
-package com.ainote.server.memo.dto;
+package com.ainote.server.dto;
 
 public record GetMemoResponse(MemoDto memo) {
 }
