+++ apps/server-spring/src/main/java/com/ainote/server/dto/HealthResponse.java
@@ -1,4 +1,4 @@
-package com.ainote.server.health;
+package com.ainote.server.dto;
 
 public record HealthResponse(String service, String status) {
