+++ apps/server-spring/src/main/java/com/ainote/server/dto/ListDeletionsResponse.java
@@ -1,0 +1,6 @@
+package com.ainote.server.dto;
+
+import java.util.List;
+
+public record ListDeletionsResponse(List<MemoTombstoneDto> deletions) {
+}
