+++ apps/server-spring/src/main/java/com/ainote/server/dto/ListMemosResponse.java
@@ -1,4 +1,4 @@
-package com.ainote.server.memo.dto;
+package com.ainote.server.dto;
 
 import java.util.List;
