+++ apps/server-spring/src/main/java/com/ainote/server/dto/MemoCreateRequest.java
@@ -1,4 +1,4 @@
-package com.ainote.server.memo.dto;
+package com.ainote.server.dto;
 
 public record MemoCreateRequest(String title, String body, String category, String color) {
 }
