--- apps/server-spring/src/main/java/com/ainote/server/memo/dto/ListDeletionsResponse.java
@@ -1,5 +1,0 @@
-package com.ainote.server.memo.dto;
-
-import java.util.List;
-
-public record ListDeletionsResponse(List<MemoTombstoneDto> deletions) {}
