+++ apps/server-spring/src/main/java/com/ainote/server/repository/MemoRepository.java
@@ -1,5 +1,6 @@
-package com.ainote.server.memo;
+package com.ainote.server.repository;
 
+import com.ainote.server.domain.Memo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemoRepository extends JpaRepository<Memo, String> {
