+++ apps/server-spring/src/test/java/com/ainote/server/controller/HealthControllerTest.java
@@ -1,8 +1,9 @@
-package com.ainote.server.health;
+package com.ainote.server.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.boot.test.context.SpringBootTest.WebEnvironment.RANDOM_PORT;
 
+import com.ainote.server.dto.HealthResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
